+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/command_attachment/gui/CommandAttachmentEditorGui.java
@@ -1,0 +1,93 @@
+package io.github.sakurawald.fuji.module.initializer.command_attachment.gui;
+
+import eu.pb4.sgui.api.elements.GuiElementBuilder;
+import eu.pb4.sgui.api.elements.GuiElementInterface;
+import eu.pb4.sgui.api.gui.SimpleGui;
+import io.github.sakurawald.fuji.core.auxiliary.minecraft.EntityHelper;
+import io.github.sakurawald.fuji.core.auxiliary.minecraft.TextHelper;
+import io.github.sakurawald.fuji.core.auxiliary.minecraft.UuidHelper;
+import io.github.sakurawald.fuji.core.auxiliary.minecraft.WorldHelper;
+import io.github.sakurawald.fuji.core.command.exception.AbortCommandExecutionException;
+import io.github.sakurawald.fuji.core.gui.component.gui.PagedGui;
+import io.github.sakurawald.fuji.module.initializer.command_attachment.service.CommandAttachmentService;
+import io.github.sakurawald.fuji.module.initializer.command_attachment.structure.CommandAttachmentDataNode;
+import io.github.sakurawald.fuji.module.initializer.command_attachment.structure.attachment_entry.BaseCommandAttachmentEntry;
+import java.util.List;
+import java.util.Optional;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.server.world.ServerWorld;
+import net.minecraft.text.Text;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public class CommandAttachmentEditorGui extends PagedGui<BaseCommandAttachmentEntry> {
+
+    private final CommandAttachmentDataNode commandAttachmentDataNode;
+
+    public CommandAttachmentEditorGui(@Nullable SimpleGui parent, @NotNull ServerPlayerEntity player, @NotNull List<BaseCommandAttachmentEntry> entities, int pageIndex, CommandAttachmentDataNode commandAttachmentDataNode) {
+        super(parent, player, TextHelper.getTextByKey(player, "command_attachment.editor.gui.title", commandAttachmentDataNode.getId()), entities, pageIndex);
+        this.commandAttachmentDataNode = commandAttachmentDataNode;
+    }
+
+    @Override
+    protected PagedGui<BaseCommandAttachmentEntry> makePage(@Nullable SimpleGui parent, @NotNull ServerPlayerEntity player, Text title, @NotNull List<BaseCommandAttachmentEntry> entities, int pageIndex) {
+        return new CommandAttachmentEditorGui(parent, player, entities, pageIndex, this.commandAttachmentDataNode);
+    }
+
+    @SuppressWarnings("OptionalIsPresent")
+    private static Optional<String> getTargetUuidForEditor(@NotNull ServerPlayerEntity player) {
+        /* Try to get UUID from looking at entity. */
+        Optional<String> lookingAtEntityUUID = WorldHelper.Raycast
+            .getLookingAtEntity(player)
+            .map(UuidHelper::getAttachedUuid);
+        if (lookingAtEntityUUID.isPresent()) {
+            return lookingAtEntityUUID;
+        }
+
+        /* Try to get UUID from looking at block. */
+        ServerWorld world = EntityHelper.getServerWorld(player);
+        Optional<String> lookingAtBlockUUID = WorldHelper.Raycast
+            .getLookingAtBlock(player)
+            .map(block -> UuidHelper.getAttachedUuid(world, block))
+            .filter(uuid -> CommandAttachmentService.findAttachmentDataNode(uuid).isPresent());
+        if (lookingAtBlockUUID.isPresent()) {
+            return lookingAtBlockUUID;
+        }
+
+        /* Try to get UUID from main hand item. */
+        ItemStack mainHandStack = player.getMainHandStack();
+        Optional<String> mainHandItemUUID = UuidHelper.getAttachedUuid(mainHandStack);
+        if (mainHandItemUUID.isPresent()) {
+            return mainHandItemUUID;
+        }
+
+        /* Failed to get the UUID. */
+        return Optional.empty();
+    }
+
+    public static @NotNull CommandAttachmentEditorGui make(@NotNull ServerPlayerEntity player) {
+        return getTargetUuidForEditor(player)
+            .flatMap(CommandAttachmentService::findAttachmentDataNode)
+            .map(node -> {
+                List<BaseCommandAttachmentEntry> entities = node.getAttachments().getEntries();
+                return new CommandAttachmentEditorGui(null, player, entities, 0, node);
+            })
+            .orElseThrow(() -> {
+                TextHelper.sendTextByKey(player, "command_attachment.editor.no_target");
+                return new AbortCommandExecutionException();
+            });
+    }
+
+    @Override
+    protected @NotNull GuiElementInterface toGuiElement(@NotNull BaseCommandAttachmentEntry entity) {
+        GuiElementBuilder builder = new GuiElementBuilder();
+
+        builder.setItem(Items.NAME_TAG);
+        builder.setName(TextHelper.getTextByKey(player, "command_attachment.attachment"));
+        builder.setLore(entity.asLore(player));
+
+        return builder.build();
+    }
+}